+++ internal/data/ent/schema/user_favor.go
@@ -20,7 +20,7 @@ func (UserFavor) Fields() []ent.Field {
 	return []ent.Field{
 		field.Int64("user_id").Comment(""),
 		field.Int64("spu_id").Comment(""),
-		field.Int8("status").Default(1).Comment("1收藏 0取消收藏"),
+		field.Int("status").Default(1).Comment("1收藏 0取消收藏"),
 	}
 }
 func (UserFavor) Mixin() []ent.Mixin {
