+++ internal/data/model/order.go
@@ -38,7 +38,7 @@ type Order struct {
 	// FinalTotalPrice holds the value of the "final_total_price" field.
 	FinalTotalPrice float64 `json:"final_total_price,omitempty"`
 	// Status holds the value of the "status" field.
-	Status int8 `json:"status,omitempty"`
+	Status int `json:"status,omitempty"`
 	// Edges holds the relations/edges for other nodes in the graph.
 	// The values are being populated by the OrderQuery when eager-loading is set.
 	Edges OrderEdges `json:"edges"`
@@ -181,7 +181,7 @@ func (o *Order) assignValues(columns []string, values []interface{}) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field status", values[i])
 			} else if value.Valid {
-				o.Status = int8(value.Int64)
+				o.Status = int(value.Int64)
 			}
 		}
 	}
