+++ internal/data/model/ordersub/where.go
@@ -150,7 +150,7 @@ func FinalPrice(v float64) predicate.OrderSub {
 }
 
 // Status applies equality check predicate on the "status" field. It's identical to StatusEQ.
-func Status(v int8) predicate.OrderSub {
+func Status(v int) predicate.OrderSub {
 	return predicate.OrderSub(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldStatus), v))
 	})
@@ -835,21 +835,21 @@ func FinalPriceLTE(v float64) predicate.OrderSub {
 }
 
 // StatusEQ applies the EQ predicate on the "status" field.
-func StatusEQ(v int8) predicate.OrderSub {
+func StatusEQ(v int) predicate.OrderSub {
 	return predicate.OrderSub(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldStatus), v))
 	})
 }
 
 // StatusNEQ applies the NEQ predicate on the "status" field.
-func StatusNEQ(v int8) predicate.OrderSub {
+func StatusNEQ(v int) predicate.OrderSub {
 	return predicate.OrderSub(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldStatus), v))
 	})
 }
 
 // StatusIn applies the In predicate on the "status" field.
-func StatusIn(vs ...int8) predicate.OrderSub {
+func StatusIn(vs ...int) predicate.OrderSub {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -866,7 +866,7 @@ func StatusIn(vs ...int8) predicate.OrderSub {
 }
 
 // StatusNotIn applies the NotIn predicate on the "status" field.
-func StatusNotIn(vs ...int8) predicate.OrderSub {
+func StatusNotIn(vs ...int) predicate.OrderSub {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -883,28 +883,28 @@ func StatusNotIn(vs ...int8) predicate.OrderSub {
 }
 
 // StatusGT applies the GT predicate on the "status" field.
-func StatusGT(v int8) predicate.OrderSub {
+func StatusGT(v int) predicate.OrderSub {
 	return predicate.OrderSub(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldStatus), v))
 	})
 }
 
 // StatusGTE applies the GTE predicate on the "status" field.
-func StatusGTE(v int8) predicate.OrderSub {
+func StatusGTE(v int) predicate.OrderSub {
 	return predicate.OrderSub(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldStatus), v))
 	})
 }
 
 // StatusLT applies the LT predicate on the "status" field.
-func StatusLT(v int8) predicate.OrderSub {
+func StatusLT(v int) predicate.OrderSub {
 	return predicate.OrderSub(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldStatus), v))
 	})
 }
 
 // StatusLTE applies the LTE predicate on the "status" field.
-func StatusLTE(v int8) predicate.OrderSub {
+func StatusLTE(v int) predicate.OrderSub {
 	return predicate.OrderSub(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldStatus), v))
 	})
